requirements.txt
streamlit>=1.37
pydantic>=2.7
numpy>=1.26
pandas>=2.2
scipy>=1.13
scikit-image>=0.24
pillow>=10.4
opencv-python-headless>=4.10
monai>=1.3
nibabel>=5.2
simpleitk>=2.3
torch>=2.2; platform_system != "Darwin"
torchvision>=0.17; platform_system != "Darwin"
python-dotenv>=1.0
tqdm>=4.66
openai>=1.45
torchxrayvision
google-generativeai
xgboost
ramanspy
matplotlib
pydantic
transformers
tensorflow